use subprocess
